Move relevant components to DandisetLandingView/
